refactor(cli): extract report rendering helpers
