+++ skylos/cli.py
@@ -26,7 +26,11 @@ from skylos.constants import (
     DEFAULT_EXCLUDE_FOLDERS,
     get_non_library_dir_kind,
 )
-from skylos.config import ConfigError, load_config, resolve_config_file_path
+from skylos.config import (
+    ConfigError as ConfigError,
+    load_config,
+    resolve_config_file_path,
+)
 from skylos.cloud.credentials import PROVIDERS
 from skylos.core.result_cache import (
     build_trace_cache_key,
@@ -40,10 +44,52 @@ from skylos.remediation.safety import resolve_remediation_path
 from pathlib import Path
 import pathlib
 import skylos
-from collections import defaultdict
 from io import StringIO
 import subprocess
 import textwrap
+
+from skylos.reporting.github_annotations import (
+    _emit_github_annotation as _emit_github_annotation,
+    _emit_github_annotations,
+    _emit_github_grade_annotation as _emit_github_grade_annotation,
+    _filter_github_annotations_by_severity as _filter_github_annotations_by_severity,
+    _github_annotation_items as _github_annotation_items,
+    _github_annotation_sort_key as _github_annotation_sort_key,
+    _github_dead_code_annotation as _github_dead_code_annotation,
+    _github_finding_annotation as _github_finding_annotation,
+)
+from skylos.reporting.llm_report import (
+    _collect_llm_report_findings as _collect_llm_report_findings,
+    _default_dead_code_llm_fields as _default_dead_code_llm_fields,
+    _format_llm_report_section as _format_llm_report_section,
+    _generate_llm_report as _generate_llm_report,
+    _llm_report_code_block as _llm_report_code_block,
+    _llm_report_secret_block as _llm_report_secret_block,
+    _llm_report_sort_key as _llm_report_sort_key,
+)
+from skylos.ui.rich_report import (
+    _dead_code_why as _dead_code_why,
+    _display_cap as _display_cap,
+    _display_rule_name as _display_rule_name,
+    _format_confidence as _format_confidence,
+    _grep_verify_pill as _grep_verify_pill,
+    _quality_detail as _quality_detail,
+    _render_circular_deps as _render_circular_deps,
+    _render_custom_rules as _render_custom_rules,
+    _render_danger as _render_danger,
+    _render_grade as _render_grade,
+    _render_quality as _render_quality,
+    _render_result_tree as _render_result_tree,
+    _render_sca as _render_sca,
+    _render_secrets as _render_secrets,
+    _render_unused as _render_unused,
+    _render_unused_simple as _render_unused_simple,
+    _results_pill as _results_pill,
+    _score_style as _score_style,
+    _verification_label as _verification_label,
+    _verification_proof as _verification_proof,
+    render_results,
+)
 
 from rich.console import Console
 from rich.table import Table
@@ -54,8 +100,6 @@ from rich.progress import TextColumn as TextColumn
 from rich.theme import Theme
 from rich.logging import RichHandler
 from rich.rule import Rule
-from rich.tree import Tree
-from rich.markup import escape
 
 class _LazyInquirer:
     """Import inquirer only when an interactive prompt is actually used."""
@@ -1282,261 +1326,6 @@ def print_badge(
     console.print("```")
 
 
-_LLM_REPORT_CATEGORIES = [
-    ("danger", "Security"),
-    ("secrets", "Secrets"),
-    ("quality", "Quality"),
-    ("custom_rules", "Custom Rules"),
-]
-_LLM_REPORT_DEAD_CODE_META = {
-    "unused_functions": ("SKY-DC001", "MEDIUM", "Unused function"),
-    "unused_imports": ("SKY-DC002", "LOW", "Unused import"),
-    "unused_classes": ("SKY-DC003", "MEDIUM", "Unused class"),
-    "unused_variables": ("SKY-DC004", "LOW", "Unused variable"),
-    "unused_parameters": ("SKY-DC005", "LOW", "Unused parameter"),
-    "unused_files": ("SKY-DC006", "LOW", "Empty file"),
-}
-_LLM_REPORT_SEVERITY_ORDER = {"CRITICAL": 0, "HIGH": 1, "MEDIUM": 2, "LOW": 3}
-
-
-def _default_dead_code_llm_fields(finding, rule_id, severity, human_label):
-    if not finding.get("message"):
-        name = finding.get("name") or finding.get("simple_name") or ""
-        why = finding.get("why_unused")
-        if why:
-            finding["message"] = (
-                f"{human_label} '{name}' is never used ({', '.join(why)})"
-            )
-        else:
-            finding["message"] = f"{human_label} '{name}' is never used"
-    if not finding.get("rule_id"):
-        finding["rule_id"] = rule_id
-    if not finding.get("severity"):
-        finding["severity"] = severity
-
-
-def _collect_llm_report_findings(result: dict):
-    all_findings = []
-    for category, label in _LLM_REPORT_CATEGORIES:
-        for finding in result.get(category, []):
-            all_findings.append((finding, label))
-
-    for category in _LLM_REPORT_DEAD_CODE_META:
-        rule_id, severity, human_label = _LLM_REPORT_DEAD_CODE_META[category]
-        for finding in result.get(category, []):
-            _default_dead_code_llm_fields(finding, rule_id, severity, human_label)
-            all_findings.append((finding, "Dead Code"))
-
-    return all_findings
-
-
-def _llm_report_sort_key(finding_with_label):
-    finding, _label = finding_with_label
-    return _LLM_REPORT_SEVERITY_ORDER.get(finding.get("severity", "LOW"), 4)
-
-
-def _llm_report_code_block(
-    file_path: str, line: int, project_root: pathlib.Path, file_cache: dict
-) -> str:
-    if not file_path:
-        return ""
-    try:
-        line_number = int(line)
-    except (TypeError, ValueError):
-        return ""
-    if line_number < 1:
-        return ""
-
-    abs_path = pathlib.Path(file_path)
-    if not abs_path.is_absolute():
-        abs_path = project_root / abs_path
-    cache_key = str(abs_path)
-
-    if cache_key not in file_cache:
-        try:
-            if abs_path.is_file():
-                file_cache[cache_key] = abs_path.read_text(
-                    encoding="utf-8", errors="replace"
-                ).splitlines()
-            else:
-                file_cache[cache_key] = None
-        except (OSError, ValueError) as exc:
-            logging.getLogger(__name__).debug(
-                "Failed to read LLM report context from %s: %s", abs_path, exc
-            )
-            file_cache[cache_key] = None
-
-    src_lines = file_cache[cache_key]
-    if src_lines is not None:
-        start = max(0, line_number - 3)
-        end = min(len(src_lines), line_number + 4)
-        context_lines = []
-        for i in range(start, end):
-            marker = ">>>" if i == line_number - 1 else "   "
-            context_lines.append(f"{marker} {i + 1:4d} | {src_lines[i]}")
-        if context_lines:
-            return "\n```\n" + "\n".join(context_lines) + "\n```\n"
-    return ""
-
-
-def _llm_report_secret_block(finding: dict) -> str:
-    preview = finding.get("preview") or "****"
-    return f"\n```\n>>> secret preview | {preview}\n```\n"
-
-
-def _format_llm_report_section(finding_num, finding, label, code_block):
-    rule_id = finding.get("rule_id", "")
-    severity = finding.get("severity", "INFO")
-    name = finding.get("name") or finding.get("simple_name", "")
-    file_path = finding.get("file", "")
-    line = finding.get("line", 0)
-    message = finding.get("message", "")
-
-    return (
-        f"\n## {finding_num}. {rule_id} | {severity} | {label}\n"
-        f"File: {file_path}:{line}\n"
-        f"Name: {name}\n"
-        f"{code_block}\n"
-        f"Problem: {message}\n"
-        f"\n---\n"
-    )
-
-
-def _generate_llm_report(result: dict, project_root: pathlib.Path) -> str:
-    all_findings = _collect_llm_report_findings(result)
-    if not all_findings:
-        return "# Skylos Report\n\nNo findings.\n"
-
-    all_findings.sort(key=_llm_report_sort_key)
-
-    sections = [
-        f"# Skylos Report — {len(all_findings)} findings\n\n"
-        f"Fix each finding below. The code context shows the problematic lines.\n\n---\n"
-    ]
-    file_cache = {}
-
-    for finding_num, (finding, label) in enumerate(all_findings, 1):
-        file_path = finding.get("file", "")
-        line = finding.get("line", 0)
-        code_block = (
-            _llm_report_secret_block(finding)
-            if label == "Secrets"
-            else _llm_report_code_block(file_path, line, project_root, file_cache)
-        )
-        sections.append(
-            _format_llm_report_section(finding_num, finding, label, code_block)
-        )
-
-    return "".join(sections)
-
-
-_GITHUB_ANNOTATION_LEVELS = {
-    "CRITICAL": "error",
-    "HIGH": "error",
-    "MEDIUM": "warning",
-    "LOW": "notice",
-}
-_GITHUB_ANNOTATION_PRIORITY = {"CRITICAL": 0, "HIGH": 1, "MEDIUM": 2, "LOW": 3}
-_GITHUB_ANNOTATION_THRESHOLDS = {
-    "critical": {"CRITICAL"},
-    "high": {"CRITICAL", "HIGH"},
-    "medium": {"CRITICAL", "HIGH", "MEDIUM"},
-    "low": {"CRITICAL", "HIGH", "MEDIUM", "LOW"},
-}
-_GITHUB_FINDING_CATEGORIES = ("danger", "quality", "secrets", "custom_rules")
-_GITHUB_DEAD_CODE_CATEGORIES = (
-    ("unused_functions", "Unused function"),
-    ("unused_imports", "Unused import"),
-    ("unused_classes", "Unused class"),
-    ("unused_variables", "Unused variable"),
-    ("unused_parameters", "Unused parameter"),
-)
-
-
-def _emit_github_grade_annotation(result):
-    grade_data = result.get("grade")
-    if grade_data:
-        overall = grade_data["overall"]
-        print(
-            f"::notice title=Skylos Grade::{overall['letter']} ({overall['score']}/100)"
-        )
-
-
-def _github_finding_annotation(finding):
-    file = finding.get("file") or finding.get("file_path") or ""
-    line = finding.get("line") or finding.get("line_number") or 1
-    msg = (
-        finding.get("message")
-        or finding.get("msg")
-        or finding.get("detail")
-        or "Issue detected"
-    )
-    rule_id = finding.get("rule_id") or ""
-    severity = finding.get("severity", "MEDIUM").upper()
-    title = f"Skylos {rule_id}" if rule_id else "Skylos"
-    return {
-        "file": file,
-        "line": line,
-        "msg": msg,
-        "title": title,
-        "severity": severity,
-    }
-
-
-def _github_dead_code_annotation(item, label):
-    name = item.get("name", "") if isinstance(item, dict) else str(item)
-    file = item.get("file", "") if isinstance(item, dict) else ""
-    line = item.get("line", 1) if isinstance(item, dict) else 1
-    return {
-        "file": file,
-        "line": line,
-        "msg": f"{label}: {name}",
-        "title": "Skylos Dead Code",
-        "severity": "MEDIUM",
-    }
-
-
-def _github_annotation_items(result):
-    annotations = []
-    for category in _GITHUB_FINDING_CATEGORIES:
-        for finding in result.get(category, []) or []:
-            annotations.append(_github_finding_annotation(finding))
-
-    for category, label in _GITHUB_DEAD_CODE_CATEGORIES:
-        for item in result.get(category, []) or []:
-            annotations.append(_github_dead_code_annotation(item, label))
-    return annotations
-
-
-def _filter_github_annotations_by_severity(annotations, severity_filter):
-    if severity_filter:
-        allowed = _GITHUB_ANNOTATION_THRESHOLDS.get(severity_filter, set())
-        return [a for a in annotations if a["severity"] in allowed]
-    return annotations
-
-
-def _github_annotation_sort_key(annotation):
-    return _GITHUB_ANNOTATION_PRIORITY.get(annotation["severity"], 99)
-
-
-def _emit_github_annotation(annotation):
-    level = _GITHUB_ANNOTATION_LEVELS.get(annotation["severity"], "warning")
-    print(
-        f"::{level} file={annotation['file']},line={annotation['line']},"
-        f"title={annotation['title']}::{annotation['msg']}"
-    )
-
-
-def _emit_github_annotations(result, *, max_annotations=50, severity_filter=None):
-    _emit_github_grade_annotation(result)
-    annotations = _github_annotation_items(result)
-    annotations = _filter_github_annotations_by_severity(annotations, severity_filter)
-    annotations.sort(key=_github_annotation_sort_key)
-
-    for annotation in annotations[:max_annotations]:
-        _emit_github_annotation(annotation)
-
-
 _DISPLAY_FILTER_SEVERITY_RANK = {"critical": 4, "high": 3, "medium": 2, "low": 1}
 _DISPLAY_FILTER_CATEGORY_MAP = {
     "unused_functions": "dead_code",
@@ -1619,724 +1408,6 @@ def _apply_display_filters(result, severity=None, category=None, file_filter=Non
         filtered[key] = _display_filter_items(items, file_filter, min_rank)
 
     return filtered
-
-
-_RESULTS_SUPPRESS_HINT = '[muted]Suppress: # skylos: ignore (line), ignore = ["SKY-XXX"] (rule), or # skylos: ignore-start/end (block)[/muted]\n'
-_RESULTS_DOCS_LINK = (
-    _RESULTS_SUPPRESS_HINT
-    + "[muted]Full guide: https://docs.skylos.dev/guides/understanding-output[/muted]\n"
-)
-
-
-def _results_pill(label, n, ok_style="good", bad_style="bad"):
-    if n == 0:
-        style = ok_style
-    else:
-        style = bad_style
-    return f"[{style}]{label}: {n}[/{style}]"
-
-
-def _grep_verify_pill(summary):
-    grep_verify = summary.get("grep_verify")
-    if not isinstance(grep_verify, dict):
-        return None
-    if not grep_verify.get("enabled"):
-        return "[muted]Grep verify: off[/muted]"
-    rescued_count = int(grep_verify.get("rescued_count") or 0)
-    return f"[brand]Grep verify: on[/brand] [muted](rescued {rescued_count})[/muted]"
-
-
-def _display_cap(items, limit):
-    cap = limit or len(items)
-    return items[:cap], max(0, len(items) - cap)
-
-
-def _score_style(score):
-    if score >= 90:
-        return "good"
-    if score >= 80:
-        return "brand"
-    if score >= 70:
-        return "yellow"
-    return "bad"
-
-
-def _render_grade(console: Console, grade_data, *, copy_badge: bool = True):
-    from skylos.reporting.grader import generate_badge_url
-
-    overall = grade_data["overall"]
-    cats = grade_data["categories"]
-    o_score = overall["score"]
-    g_style = _score_style(o_score)
-
-    console.print(
-        Panel.fit(
-            f"[{g_style}]Codebase Grade: {overall['letter']} ({o_score}/100)[/{g_style}]",
-            border_style=g_style,
-        )
-    )
-
-    grade_table = Table(title="Grade Breakdown", expand=True)
-    grade_table.add_column("Category", style="bold", width=16)
-    grade_table.add_column("Score", justify="right", width=8)
-    grade_table.add_column("Grade", width=6)
-    grade_table.add_column("Weight", style="muted", width=8)
-    grade_table.add_column("Key Issue", overflow="fold")
-
-    default_category_order = (
-        "security",
-        "quality",
-        "dead_code",
-        "dependencies",
-        "secrets",
-    )
-    category_order = grade_data.get("scanned_categories") or default_category_order
-
-    for cat_name in category_order:
-        if cat_name not in cats:
-            continue
-        cat = cats[cat_name]
-        display_name = cat_name.replace("_", " ").title()
-        s_val = cat["score"]
-        l_val = cat["letter"]
-        w_pct = f"{int(cat['weight'] * 100)}%"
-        issue = cat.get("key_issue") or "-"
-        if len(issue) > 60:
-            issue = issue[:57] + "..."
-
-        s_style = _score_style(s_val)
-        s_str = f"[{s_style}]{s_val}[/{s_style}]"
-        l_str = f"[{s_style}]{l_val}[/{s_style}]"
-
-        grade_table.add_row(display_name, s_str, l_str, w_pct, issue)
-
-    console.print(grade_table)
-    badge_url = generate_badge_url(overall["letter"], o_score)
-    badge_markdown = (
-        f"[![Skylos Grade]({badge_url})](https://github.com/duriantaco/skylos)"
-    )
-
-    console.print()
-    console.print(
-        Panel.fit(
-            "[bold cyan]Score Badge for your README.md:[/bold cyan]\n\n"
-            f"[yellow]{badge_markdown}[/yellow]",
-            title="[cyan]Score Badge[/cyan]",
-            border_style="cyan",
-        )
-    )
-
-    if copy_badge:
-        try:
-            import pyperclip
-
-            pyperclip.copy(badge_markdown)
-            console.print("[good]Copied to clipboard![/good]")
-        except ImportError:
-            console.print(
-                "[muted]Install pyperclip for auto-copy: pip install pyperclip[/muted]"
-            )
-        except (pyperclip.PyperclipException, OSError) as exc:
-            logger.debug("Failed to copy badge markdown to clipboard: %s", exc)
-
-    console.print()
-
-
-def _format_confidence(conf):
-    if isinstance(conf, int):
-        if conf >= 90:
-            return f"[red]{conf}%[/red]"
-        if conf >= 75:
-            return f"[yellow]{conf}%[/yellow]"
-        return f"[dim]{conf}%[/dim]"
-    return str(conf)
-
-
-def _render_unused(console: Console, root_path, limit, title, items, name_key="name"):
-    if not items:
-        return
-
-    console.rule(f"[bold]{title}")
-
-    table = Table(expand=True)
-    table.add_column("#", style="muted", width=3)
-    table.add_column("Name", style="bold")
-    table.add_column("Location", style="muted", overflow="fold")
-    table.add_column("Conf", style="yellow", width=6, justify="right")
-
-    show, overflow = _display_cap(items, limit)
-    for i, item in enumerate(show, 1):
-        nm = item.get(name_key) or item.get("simple_name") or "<?>"
-        short = _shorten_path(item.get("file"), root_path)
-        loc = f"{short}:{item.get('line', '?')}"
-        conf_str = _format_confidence(item.get("confidence", "?"))
-        table.add_row(str(i), nm, loc, conf_str)
-
-    console.print(table)
-    if overflow:
-        console.print(
-            f"  [muted]... and {overflow} more (use --limit to adjust)[/muted]"
-        )
-    console.print(
-        "[muted]Name — the unused function, import, class, or variable.[/muted]\n"
-        "[muted]Conf — how confident Skylos is that this code is truly unused (higher = safer to remove).[/muted]\n"
-        + _RESULTS_DOCS_LINK
-    )
-
-
-def _render_unused_simple(
-    console: Console, root_path, limit, title, items, name_key="name"
-):
-    if not items:
-        return
-
-    console.rule(f"[bold]{title}")
-
-    table = Table(expand=True)
-    table.add_column("#", style="muted", width=3)
-    table.add_column("Name", style="bold")
-    table.add_column("Location", style="muted", overflow="fold")
-
-    show, overflow = _display_cap(items, limit)
-    for i, item in enumerate(show, 1):
-        nm = item.get(name_key) or item.get("simple_name") or "<?>"
-        short = _shorten_path(item.get("file"), root_path)
-        loc = f"{short}:{item.get('line', '?')}"
-        table.add_row(str(i), nm, loc)
-
-    console.print(table)
-    if overflow:
-        console.print(
-            f"  [muted]... and {overflow} more (use --limit to adjust)[/muted]"
-        )
-    console.print()
-
-
-def _quality_detail(quality):
-    raw_kind = quality.get("kind") or quality.get("metric") or "quality"
-    func = quality.get("name") or quality.get("simple_name") or "<?>"
-    value = quality.get("value") or quality.get("complexity")
-    thr = quality.get("threshold")
-    length = quality.get("length")
-    qtype = quality.get("type", "")
-
-    if qtype == "string":
-        detail = f"repeated {value}×"
-        if thr is not None:
-            detail += f" (max {thr})"
-        func = f'"{func}"'
-    elif qtype == "dependency":
-        detail = str(value)
-    elif raw_kind in {
-        "typing",
-        "framework",
-        "framework_security",
-        "repo_policy",
-    }:
-        detail = quality.get("message") or str(value)
-    elif raw_kind == "nesting":
-        detail = f"Deep nesting: depth {value}"
-    elif raw_kind == "structure":
-        detail = f"Line count: {value}"
-    elif raw_kind == "complexity":
-        detail = f"Complexity: {value}"
-        if thr is not None:
-            detail += f" (max {thr})"
-    else:
-        detail = f"{value}"
-        if thr is not None:
-            detail += f" (max {thr})"
-    if length is not None:
-        detail += f", {length} lines"
-
-    return raw_kind.replace("_", " ").title(), func, detail
-
-
-def _render_quality(console: Console, limit, items):
-    if not items:
-        return
-
-    console.rule("[bold red]Quality Issues")
-    table = Table(expand=True)
-    table.add_column("#", style="muted", width=3)
-    table.add_column("Type", style="yellow", width=12)
-    table.add_column("Name", style="bold")
-    table.add_column("Detail")
-    table.add_column("Location", style="muted", width=36)
-
-    show, overflow = _display_cap(items, limit)
-    for i, quality in enumerate(show, 1):
-        kind, func, detail = _quality_detail(quality)
-        loc = f"{quality.get('basename', '?')}:{quality.get('line', '?')}"
-        table.add_row(str(i), escape(kind), escape(func), escape(detail), escape(loc))
-
-    console.print(table)
-    if overflow:
-        console.print(
-            f"  [muted]... and {overflow} more (use --limit to adjust)[/muted]"
-        )
-    console.print(
-        "[muted]Reading the table:[/muted]\n"
-        "[muted]  • Complexity — number of branches/loops in a function (lower = easier to test)[/muted]\n"
-        "[muted]  • Nesting — how deeply indented the code is (depth count)[/muted]\n"
-        "[muted]  • Structure — line count of a function or argument count[/muted]\n"
-        "[muted]  • Duplicate strings — how many times a literal appears[/muted]\n"
-        '[muted]  • "max N" / "(max N)" — the configured threshold; tune in [tool.skylos] (complexity, nesting, max_args, max_lines, duplicate_strings)[/muted]\n'
-        + _RESULTS_DOCS_LINK
-    )
-
-
-def _render_circular_deps(console: Console, limit, items):
-    if not items:
-        return
-
-    console.rule("[bold yellow]Circular Dependencies")
-    table = Table(expand=True)
-    table.add_column("#", style="muted", width=3)
-    table.add_column("Cycle", style="bold")
-    table.add_column("Length", width=6)
-    table.add_column("Severity", width=8)
-    table.add_column("Suggested Break", style="cyan")
-
-    show, overflow = _display_cap(items, limit)
-    for i, cd in enumerate(show, 1):
-        cycle = cd.get("cycle", [])
-        cycle_str = " → ".join(cycle) + f" → {cycle[0]}" if cycle else "?"
-        length = str(cd.get("cycle_length", len(cycle)))
-        sev = cd.get("severity", "MEDIUM")
-        suggested = cd.get("suggested_break", "?")
-        table.add_row(str(i), cycle_str, length, sev, suggested)
-
-    console.print(table)
-    if overflow:
-        console.print(
-            f"  [muted]... and {overflow} more (use --limit to adjust)[/muted]"
-        )
-    console.print(
-        "[muted]Cycle — the chain of modules that import each other in a loop.[/muted]\n"
-        "[muted]Length — how many modules are in the cycle.[/muted]\n"
-        "[muted]Suggested Break — the module to refactor to break the dependency loop.[/muted]\n"
-        + _RESULTS_DOCS_LINK
-    )
-
-
-def _render_custom_rules(console: Console, root_path, limit, items):
-    custom = [
-        i for i in (items or []) if str(i.get("rule_id", "")).startswith("CUSTOM-")
-    ]
-    if not custom:
-        return
-
-    console.rule("[bold magenta]Custom Rules")
-    table = Table(expand=True)
-    table.add_column("#", style="muted", width=3)
-    table.add_column("Rule", style="magenta", width=18)
-    table.add_column("Severity", width=10)
-    table.add_column("Message", overflow="fold")
-    table.add_column("Location", style="muted", width=36)
-
-    show, overflow = _display_cap(custom, limit)
-    for i, d in enumerate(show, 1):
-        rule = d.get("rule_id") or "CUSTOM"
-        sev = d.get("severity") or "MEDIUM"
-        msg = d.get("message") or "Custom rule violation"
-        short = _shorten_path(d.get("file"), root_path)
-        loc = f"{short}:{d.get('line', '?')}"
-        table.add_row(str(i), rule, sev, msg, loc)
-
-    console.print(table)
-    if overflow:
-        console.print(
-            f"  [muted]... and {overflow} more (use --limit to adjust)[/muted]"
-        )
-    console.print()
-
-
-def _render_secrets(console: Console, root_path, limit, items):
-    if not items:
-        return
-
-    console.rule("[bold red]Secrets")
-    has_provenance = any(s.get("ai_authored") is not None for s in (items or []))
-
-    table = Table(expand=True)
-    table.add_column("#", style="muted", width=3)
-    table.add_column("Provider", style="yellow", width=14)
-    table.add_column("Message")
-    table.add_column("Preview", style="muted", width=18)
-    table.add_column("Location", style="muted", overflow="fold")
-
-    if has_provenance:
-        table.add_column("AI", width=12)
-
-    show, overflow = _display_cap(items, limit)
-    for i, s in enumerate(show, 1):
-        prov = s.get("provider") or "generic"
-        msg = s.get("message") or "Secret detected"
-        prev = s.get("preview") or "****"
-        short = _shorten_path(s.get("file"), root_path)
-        loc = f"{short}:{s.get('line', '?')}"
-        row = [str(i), prov, msg, prev, loc]
-
-        if has_provenance:
-            if s.get("ai_authored"):
-                agent = s.get("ai_agent") or "ai"
-                row.append(f"[red]{agent}[/red]")
-            else:
-                row.append("[muted]-[/muted]")
-
-        table.add_row(*row)
-
-    console.print(table)
-    if overflow:
-        console.print(
-            f"  [muted]... and {overflow} more (use --limit to adjust)[/muted]"
-        )
-    console.print(
-        '[muted]Provider — the service the secret belongs to (e.g. AWS, Stripe, GitHub) or "generic" for high-entropy strings.[/muted]\n'
-        "[muted]Preview — a masked snippet of the detected secret.[/muted]\n"
-        + _RESULTS_DOCS_LINK
-    )
-
-
-def _render_result_tree(console: Console, result, root_path=None):
-    by_file = defaultdict(list)
-
-    def _add_unused(items, kind):
-        for u in items or []:
-            file = u.get("file")
-            if not file:
-                continue
-            line = u.get("line") or u.get("lineno") or 1
-            name = u.get("name") or u.get("simple_name") or "<?>"
-            msg = f"Unused {kind}: {name}"
-            by_file[file].append((line, "info", msg))
-
-    def _add_findings(items, kind, default_sev="medium"):
-        for f in items or []:
-            file = f.get("file")
-            if not file:
-                continue
-            line = f.get("line") or 1
-            sev = (f.get("severity") or default_sev).lower()
-            rule = f.get("rule_id")
-            msg = f.get("message") or kind
-            if rule:
-                msg = f"[{rule}] {msg}"
-            by_file[file].append((line, sev, msg))
-
-    _add_unused(result.get("unused_functions"), "function")
-    _add_unused(result.get("unused_imports"), "import")
-    _add_unused(result.get("unused_classes"), "class")
-    _add_unused(result.get("unused_variables"), "variable")
-    _add_unused(result.get("unused_parameters"), "parameter")
-
-    _add_findings(result.get("danger"), "security", default_sev="high")
-    _add_findings(result.get("secrets"), "secret", default_sev="high")
-    _add_findings(result.get("quality"), "quality", default_sev="medium")
-    _add_findings(
-        result.get("dependency_vulnerabilities"),
-        "vulnerability",
-        default_sev="high",
-    )
-
-    if not by_file:
-        console.print("[good]No findings to display.[/good]")
-        return
-
-    root_label = str(root_path) if root_path is not None else "Skylos results"
-    tree = Tree(f"[brand]{root_label}[/brand]")
-
-    for file in sorted(by_file.keys()):
-        short = _shorten_path(file, root_path)
-        file_node = tree.add(f"[bold]{short}[/bold]")
-
-        for line, sev, msg in sorted(by_file[file], key=lambda t: t[0]):
-            if sev == "high" or sev == "critical":
-                style = "bad"
-            elif sev == "medium":
-                style = "warn"
-            else:
-                style = "muted"
-            file_node.add(f"[{style}]L{line}[/{style}] {msg}")
-
-    console.print(tree)
-
-
-def _display_rule_name(rule_id):
-    from skylos.rules.catalog import get_rule_name
-
-    return get_rule_name(rule_id)
-
-
-def _verification_proof(danger_finding):
-    verification = danger_finding.get("verification")
-    if verification is None:
-        verification = {}
-
-    evidence = verification.get("evidence")
-    if evidence is None:
-        evidence = {}
-
-    chain = evidence.get("chain")
-    if isinstance(chain, list) and len(chain) > 0:
-        names = []
-        for x in chain[:6]:
-            fn = None
-            if isinstance(x, dict):
-                fn = x.get("fn")
-            if not fn:
-                fn = "?"
-            names.append(fn)
-        return " -> ".join(names)
-
-    entrypoints = evidence.get("entrypoints")
-    if entrypoints:
-        return str(len(entrypoints)) + " entrypoints scanned"
-
-    ver = verification.get("verdict")
-    if ver:
-        return "No evidence attached"
-    return ""
-
-
-def _verification_label(verdict):
-    if verdict == "VERIFIED":
-        return "[good]VERIFIED[/good]"
-    if verdict == "REFUTED":
-        return "[muted]REFUTED[/muted]"
-    if verdict == "UNKNOWN":
-        return "[warn]UNKNOWN[/warn]"
-    return "-"
-
-
-def _render_danger(console: Console, root_path, limit, items):
-    if not items:
-        return
-
-    console.rule("[bold red]Security Issues")
-
-    has_verification = any(
-        isinstance(d.get("verification"), dict) and d["verification"].get("verdict")
-        for d in (items or [])
-    )
-    has_provenance = any(d.get("ai_authored") is not None for d in (items or []))
-
-    table = Table(expand=True)
-    table.add_column("#", style="muted", width=3)
-    table.add_column("Issue", style="yellow", width=20)
-    table.add_column("Severity", width=9)
-    table.add_column("Message", overflow="fold")
-    table.add_column("Location", style="muted", width=20, overflow="fold")
-    table.add_column("Symbol", style="muted", width=10, overflow="fold")
-
-    if has_provenance:
-        table.add_column("AI", width=12)
-
-    if has_verification:
-        table.add_column("Verified", width=9)
-        table.add_column("Proof", overflow="fold")
-
-    show, overflow = _display_cap(items, limit)
-    for i, d in enumerate(show, 1):
-        rule_id = d.get("rule_id") or "UNKNOWN"
-        issue_name = _display_rule_name(rule_id)
-        issue_cell = f"{issue_name}\n[dim]{rule_id}[/dim]"
-        sev = (d.get("severity") or "UNKNOWN").title()
-        msg = d.get("message") or "Issue detected"
-        short = _shorten_path(d.get("file"), root_path)
-        loc = f"{short}:{d.get('line', '?')}"
-        symbol = d.get("symbol") or "<module>"
-        row = [str(i), issue_cell, sev, msg, loc, symbol]
-
-        if has_provenance:
-            if d.get("ai_authored"):
-                agent = d.get("ai_agent") or "ai"
-                row.append(f"[red]{agent}[/red]")
-            else:
-                row.append("[muted]-[/muted]")
-
-        if has_verification:
-            ver = (d.get("verification") or {}).get("verdict")
-            row.extend([_verification_label(ver), _verification_proof(d)])
-
-        table.add_row(*row)
-
-    console.print(table)
-    if overflow:
-        console.print(
-            f"  [muted]... and {overflow} more (use --limit to adjust)[/muted]"
-        )
-    console.print(
-        "[muted]Issue — the type of vulnerability (e.g. SQL injection, command injection, eval).[/muted]\n"
-        "[muted]Severity — risk level: Critical > High > Medium > Low.[/muted]\n"
-        "[muted]Symbol — the function or scope where the issue was found.[/muted]\n"
-        + _RESULTS_DOCS_LINK
-    )
-
-
-def _render_sca(console: Console, limit, items):
-    if not items:
-        return
-
-    console.rule("[bold red]Dependency Vulnerabilities (SCA)")
-    table = Table(expand=True)
-    table.add_column("#", style="muted", width=3)
-    table.add_column("Package", style="yellow", width=22)
-    table.add_column("Vuln ID", width=18)
-    table.add_column("Severity", width=9)
-    table.add_column("Reachability", width=14)
-    table.add_column("Message", overflow="fold")
-    table.add_column("Fix", style="good", width=14, overflow="fold")
-
-    show, overflow = _display_cap(items, limit)
-    for i, v in enumerate(show, 1):
-        meta = v.get("metadata") or {}
-        pkg = f"{meta.get('package_name', '?')}@{meta.get('package_version', '?')}"
-        vuln_id = meta.get("display_id") or meta.get("vuln_id") or v.get("rule_id", "")
-        sev = (v.get("severity") or "MEDIUM").title()
-        msg = v.get("message") or "Known vulnerability"
-        fix = meta.get("fixed_version") or "-"
-        rv = meta.get("reachability_verdict", "")
-        if rv == "reachable":
-            reach = "[red]Reachable[/red]"
-        elif rv.startswith("unreachable"):
-            reach = "[green]Unreachable[/green]"
-        elif rv == "inconclusive":
-            reach = "[yellow]Inconclusive[/yellow]"
-        else:
-            reach = "[dim]-[/dim]"
-        table.add_row(str(i), pkg, vuln_id, sev, reach, msg, fix)
-
-    console.print(table)
-    if overflow:
-        console.print(
-            f"  [muted]... and {overflow} more (use --limit to adjust)[/muted]"
-        )
-    console.print(
-        "[muted]Package — the dependency and its installed version.[/muted]\n"
-        "[muted]Reachability — whether your code actually calls the vulnerable code path.[/muted]\n"
-        "[muted]Fix — the version that patches the vulnerability (upgrade to this).[/muted]\n"
-        + _RESULTS_DOCS_LINK
-    )
-
-
-def render_results(
-    console: Console,
-    result,
-    tree=False,
-    root_path=None,
-    limit=None,
-    *,
-    copy_badge: bool = True,
-):
-    summ = result.get("analysis_summary", {})
-    console.print(
-        Panel.fit(
-            f"[brand]Python Static Analysis Results[/brand]\n[muted]Analyzed {summ.get('total_files', '?')} file(s)[/muted]",
-            border_style="brand",
-        )
-    )
-
-    console.print(
-        " ".join(
-            part
-            for part in [
-                _results_pill(
-                    "Unused functions", len(result.get("unused_functions", []))
-                ),
-                _results_pill("Unused imports", len(result.get("unused_imports", []))),
-                _results_pill(
-                    "Unused params", len(result.get("unused_parameters", []))
-                ),
-                _results_pill("Unused vars", len(result.get("unused_variables", []))),
-                _results_pill("Unused classes", len(result.get("unused_classes", []))),
-                _results_pill(
-                    "Quality", len(result.get("quality", []) or []), bad_style="warn"
-                ),
-                _results_pill(
-                    "Custom",
-                    len(result.get("custom_rules", []) or []),
-                    bad_style="warn",
-                ),
-                _results_pill(
-                    "Suppressed",
-                    len(result.get("suppressed", []) or []),
-                    ok_style="muted",
-                    bad_style="muted",
-                ),
-                _grep_verify_pill(summ),
-            ]
-            if part
-        )
-    )
-    console.print()
-
-    grade_data = result.get("grade")
-    if grade_data:
-        _render_grade(console, grade_data, copy_badge=copy_badge)
-
-    if tree:
-        _render_result_tree(console, result, root_path=root_path)
-    else:
-        _render_unused(
-            console,
-            root_path,
-            limit,
-            "Unused Functions",
-            result.get("unused_functions", []),
-            name_key="name",
-        )
-        _render_unused(
-            console,
-            root_path,
-            limit,
-            "Unused Imports",
-            result.get("unused_imports", []),
-            name_key="name",
-        )
-        _render_unused(
-            console,
-            root_path,
-            limit,
-            "Unused Parameters",
-            result.get("unused_parameters", []),
-            name_key="name",
-        )
-        _render_unused(
-            console,
-            root_path,
-            limit,
-            "Unused Variables",
-            result.get("unused_variables", []),
-            name_key="name",
-        )
-        _render_unused(
-            console,
-            root_path,
-            limit,
-            "Unused Classes",
-            result.get("unused_classes", []),
-            name_key="name",
-        )
-        _render_unused_simple(
-            console,
-            root_path,
-            limit,
-            "Unused Fixtures",
-            result.get("unused_fixtures", []),
-            name_key="name",
-        )
-        _render_secrets(console, root_path, limit, result.get("secrets", []) or [])
-        _render_danger(console, root_path, limit, result.get("danger", []) or [])
-        _render_quality(console, limit, result.get("quality", []) or [])
-        _render_circular_deps(
-            console, limit, result.get("circular_dependencies", []) or []
-        )
-        _render_custom_rules(
-            console, root_path, limit, result.get("custom_rules", []) or []
-        )
-        _render_sca(console, limit, result.get("dependency_vulnerabilities", []) or [])
 
 
 def render_pretty_results(
